feat: Add laravel-dompdf and implement PDF export functionality for all reports
This commit adds the "barryvdh/laravel-dompdf" package as a dependency in the composer.json file. It also introduces a new method, "exportAllReports", in the ReportController.php file, which generates a PDF containing all the reports. The PDF is created using the "Pdf" facade provided by the "barryvdh/laravel-dompdf" package. The generated PDF is then streamed to the user for download.

The changes also include updates to the web.php file to add two new routes: "floors/download-qrcodes" and "reports/pdf". These routes are used to trigger the PDF export functionality.

Lastly, the index.blade.php file in the "pages/report" directory is updated to include a new button for printing the PDF.

This commit addresses the need to export all reports in a PDF format, providing a convenient way for users to download and print the reports.